reference secrets via environment variables
we echo the value of DOCKERHUB_USERNAME and DOCKERHUB_REPOT_JAVET to
create a new Environment variable IMAGE_TAG. This is done to remove the
duplication, also because it is not possible to reference other
Environment variables within the `env` workflow key.

Reading the documentation, I have a hypothesis that GH doesn't allow for
echoing the contents of a secret context. Their own documentation
suggests storing the secret as an environment variable and accessing it
as such.