Infra - Redesign Docker Build (#215)
* Infra/redesign docker build (#200)

* speed up linux docker build

* update docs with command to locally build Javet artifact

* remove the GH workflow trigger on push to PR branch

* use secrets for docker hub username and repo

* prefix x86_64- in the tag for linux x86_64 docker images

* reference secrets via environment variables

we echo the value of DOCKERHUB_USERNAME and DOCKERHUB_REPOT_JAVET to
create a new Environment variable IMAGE_TAG. This is done to remove the
duplication, also because it is not possible to reference other
Environment variables within the `env` workflow key.

Reading the documentation, I have a hypothesis that GH doesn't allow for
echoing the contents of a secret context. Their own documentation
suggests storing the secret as an environment variable and accessing it
as such.

* add workflow trigger to run on push to PR branch

* remove hardcoded reference to sjtucaocao/javet from linux Dockerfiles

* Infra/redesign docker build (#200)

* speed up linux docker build

* update docs with command to locally build Javet artifact

* remove the GH workflow trigger on push to PR branch

* use secrets for docker hub username and repo

* prefix x86_64- in the tag for linux x86_64 docker images

* reference secrets via environment variables

we echo the value of DOCKERHUB_USERNAME and DOCKERHUB_REPOT_JAVET to
create a new Environment variable IMAGE_TAG. This is done to remove the
duplication, also because it is not possible to reference other
Environment variables within the `env` workflow key.

Reading the documentation, I have a hypothesis that GH doesn't allow for
echoing the contents of a secret context. Their own documentation
suggests storing the secret as an environment variable and accessing it
as such.

* add workflow trigger to run on push to PR branch

* remove hardcoded reference to sjtucaocao/javet from linux Dockerfiles

* ✨ feat(Build): Add change_node_v8_version.py

* 🐞 fix(Workflow): Tweak env vars

* 🐞 fix(Workflow): Tweak env vars v2

* 🐞 fix(Workflow): Tweak env vars v3

* 🐞 fix(Workflow): Tweak env vars v4

* 🐞 fix(Workflow): Tweak env vars v5

* 🐞 fix(Workflow): Restrict push event

* 🐞 fix(Workflow): Refactor docker file names

* 🦄 refactor(Workflow): Update badge

* 🦄 refactor(Workflow): Update naming

* 🦄 refactor(Workflow): Add stage numbers

* 🐞 fix(Workflow): Fix naming

* 🐞 fix(Workflow): Fix naming 2

* 🐞 fix(Workflow): Fix naming 3

Co-authored-by: Amith George <strider.am@gmail.com>